Fix unpredictible behaviour in index.c
When calling index_operatemailbox, index_closemailbox and index_check in a
certain order, unpredictible behaviour can happen because index_closemailbox
does not reset cache_len (neither cache_base) variable to 0.
index_check dirties the cache mmap statically known in index.c,
index_closemailbox resets cache_end, then if index_check is called again its
mmap_refresh on the cache content may do nothing (because the requested length
cache_end may not be bigger than cache_len): later on the code may thus try to
access a memory zone that have actually not been mmaped.

This can for example happen when using the IMAP commands in this order:
  SCAN, SELECT, FETCH/STORE/...
It's quite hard to reproduce because it depends on the way mmap is handled by
the kernel. Depending on the conditions, the memory zone could either be:
 - the good one: reattributed to the mmapped cyrus.cache in mailbox structure
 - an invalid one: illegal access triggered
 - a bad one: belongs to the process, but is not the cyrus.cache content