Add SQL backend failover mechanism
SQL backend has two more options: sql_backoff_time (60s by default) and
sql_idle_timeout (1h by default).

Some SQL errors (database opening failure, or connections issues upon querying
database) trigger failover to the other configured SQL hosts. In this case, the
failed host is marked unavailable for a period of sql_backoff_time and we try to
switch to one of the other available hosts.
As a special case, if failed host was inactive for at least a period of
sql_idle_timeout, we first try to reconnect to it.

Each time a query is about to be performed, we first check if we are currently
connected to the most preferred host (first one listed in configuration) and try
to switch back to a more preferred one if available.

Failover is disabled in the following cases:
  - when the query is part of an ongoing transaction
  - for the SQLite backend, since it does not use a network connection

This commit also fixes:
  - MySQL and PostgreSQL database structure not being properly closed (resource
    freeing) upon database opening failure
  - MySQL backend not logging actual error message upon database opening failure
  - MySQL backend not handling possible failure in mysql_store_result
  - SQLite database opening being considered a success even upon failure
  - compilation warning in MySQL backend